new commit

+++ src/main/java/com/example/repository/TeamRepository.java
@@ -3,7 +3,7 @@ package com.example.repository;
 import com.example.domain.Team;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TeamRepository extends JpaRepository<Team,Long>
+public interface TeamRepository extends JpaRepository<Team,Integer>
 {
 
 }
